fix(asm80): split classic instruction lowering
